refactor: address date filter review feedback
