+++ packages/schema/src/index.ts
@@ -2,4 +2,9 @@ export * from './types';
 export * from './validation';
 export * from './serializers';
 export * from './json-schema';
-export { CURRENT_SCHEMA_VERSION, migrateDashboardDefinition, isSupportedSchemaVersion } from './migrations';
+export * from './date-utils';
+export {
+  CURRENT_SCHEMA_VERSION,
+  migrateDashboardDefinition,
+  isSupportedSchemaVersion,
+} from './migrations';
